Add controllers and tests for them
Endpoint 1
for a specific Publisher it should return the list of shops selling at least one book of that publisher. Shops should be ordered by the number of books sold. Each shop should include the list of Publisher’s books that are currently in stock
Endpoint 2
For a specific Shop it should mark one or multiple copies of a book as sold.